+++ src/core/bsched.c
@@ -1531,9 +1531,8 @@ bio_write(bio_source_t *bio, gconstpointer data, size_t len)
 	amount = len > available ? available : len;
 
 	if (GNET_PROPERTY(bsched_debug) > 7)
-		g_debug("BSCHED bio_write(wio=%d, len=%lu) available=%lu",
-			bio->wio->fd(bio->wio), (unsigned long) len,
-			(unsigned long) available);
+		g_debug("BSCHED bio_write(wio=%d, len=%zu) available=%zu",
+			bio->wio->fd(bio->wio), len, available);
 
 	r = bio->wio->write(bio->wio, data, amount);
 
@@ -1546,8 +1545,8 @@ bio_write(bio_source_t *bio, gconstpointer data, size_t len)
 	 */
 
 	if ((ssize_t) -1 == r && 0 == errno) {
-		g_warning("wio->write(fd=%d, len=%lu) returned -1 with errno = 0, "
-			"assuming EAGAIN", bio->wio->fd(bio->wio), (unsigned long) len);
+		g_warning("wio->write(fd=%d, len=%zu) returned -1 with errno = 0, "
+			"assuming EAGAIN", bio->wio->fd(bio->wio), len);
 		errno = VAL_EAGAIN;
 	}
 
@@ -1644,9 +1643,8 @@ bio_writev(bio_source_t *bio, iovec_t *iov, int iovcnt)
 	 */
 
 	if (GNET_PROPERTY(bsched_debug) > 7)
-		g_debug("BSCHED bio_writev(fd=%d, len=%lu) available=%lu",
-			bio->wio->fd(bio->wio), (unsigned long) len,
-			(unsigned long) available);
+		g_debug("BSCHED bio_writev(fd=%d, len=%zu) available=%zu",
+			bio->wio->fd(bio->wio), len, available);
 
 	if (iovcnt > MAX_IOV_COUNT)
 		r = safe_writev(bio->wio, iov, iovcnt);
@@ -1662,8 +1660,8 @@ bio_writev(bio_source_t *bio, iovec_t *iov, int iovcnt)
 	 */
 
 	if ((ssize_t) -1 == r && 0 == errno) {
-		g_warning("writev(fd=%d, len=%lu) returned -1 with errno = 0, "
-			"assuming EAGAIN", bio->wio->fd(bio->wio), (unsigned long) len);
+		g_warning("writev(fd=%d, len=%zu) returned -1 with errno = 0, "
+			"assuming EAGAIN", bio->wio->fd(bio->wio), len);
 		errno = VAL_EAGAIN;
 	}
 
@@ -1721,9 +1719,8 @@ bio_sendto(bio_source_t *bio, const gnet_host_t *to,
 	}
 
 	if (GNET_PROPERTY(bsched_debug) > 7)
-		g_debug("BSCHED bio_sendto(wio=%d, len=%lu) available=%lu",
-			bio->wio->fd(bio->wio), (unsigned long) len,
-			(unsigned long) available);
+		g_debug("BSCHED bio_sendto(wio=%d, len=%zu) available=%zu",
+			bio->wio->fd(bio->wio), len, available);
 
 	g_assert(bio->wio != NULL);
 	g_assert(bio->wio->sendto != NULL);
@@ -1738,8 +1735,8 @@ bio_sendto(bio_source_t *bio, const gnet_host_t *to,
 	 */
 
 	if ((ssize_t) -1 == r && 0 == errno) {
-		g_warning("wio->sendto(fd=%d, len=%lu) returned -1 with errno = 0, "
-			"assuming EAGAIN", bio->wio->fd(bio->wio), (unsigned long) len);
+		g_warning("wio->sendto(fd=%d, len=%zu) returned -1 with errno = 0, "
+			"assuming EAGAIN", bio->wio->fd(bio->wio), len);
 		errno = VAL_EAGAIN;
 	}
 
@@ -1821,9 +1818,8 @@ bio_sendfile(sendfile_ctx_t *ctx, bio_source_t *bio,
 	amount = len > available ? available : len;
 
 	if (GNET_PROPERTY(bsched_debug) > 7)
-		g_debug("BSCHED bsched_write(fd=%d, len=%lu) available=%lu",
-			bio->wio->fd(bio->wio), (unsigned long) len,
-			(unsigned long) available);
+		g_debug("BSCHED bsched_write(fd=%d, len=%zu) available=%zu",
+			bio->wio->fd(bio->wio), len, available);
 
 #if defined(HAS_MMAP) && !defined(HAS_SENDFILE)
 	{
@@ -1949,9 +1945,9 @@ bio_sendfile(sendfile_ctx_t *ctx, bio_source_t *bio,
 
 	if (r >= 0 && *offset != start + r) {		/* Paranoid, as usual */
 		g_warning("FIXED SENDFILE returned offset: "
-			"was set to %s instead of %s (%lu byte%s written)",
+			"was set to %s instead of %s (%zu byte%s written)",
 			uint64_to_string(*offset), uint64_to_string2(start + r),
-			(unsigned long) r, r == 1 ? "" : "s");
+			r, r == 1 ? "" : "s");
 		*offset = start + r;
 	} else if ((ssize_t) -1 == r) {
 		*offset = start;	/* Paranoid: in case sendfile() touched it */
@@ -1999,9 +1995,8 @@ bio_read(bio_source_t *bio, gpointer data, size_t len)
 
 	amount = len > available ? available : len;
 	if (GNET_PROPERTY(bsched_debug) > 7)
-		g_debug("BSCHED bsched_read(fd=%d, len=%lu) available=%lu",
-			bio->wio->fd(bio->wio), (unsigned long) len,
-			(unsigned long) available);
+		g_debug("BSCHED bsched_read(fd=%d, len=%zu) available=%zu",
+			bio->wio->fd(bio->wio), len, available);
 
 	r = bio->wio->read(bio->wio, data, amount);
 	if (r > 0) {
@@ -2100,9 +2095,8 @@ bio_readv(bio_source_t *bio, iovec_t *iov, int iovcnt)
 	 */
 
 	if (GNET_PROPERTY(bsched_debug) > 7)
-		g_debug("BSCHED bio_readv(fd=%d, len=%lu) available=%lu",
-			bio->wio->fd(bio->wio), (unsigned long) len,
-			(unsigned long) available);
+		g_debug("BSCHED bio_readv(fd=%d, len=%zu) available=%zu",
+			bio->wio->fd(bio->wio), len, available);
 
 	if (iovcnt > MAX_IOV_COUNT)
 		r = safe_readv(bio->wio, iov, iovcnt);
@@ -2118,8 +2112,8 @@ bio_readv(bio_source_t *bio, iovec_t *iov, int iovcnt)
 	 */
 
 	if ((ssize_t) -1 == r && 0 == errno) {
-		g_warning("readv(fd=%d, len=%lu) returned -1 with errno = 0, "
-			"assuming EAGAIN", bio->wio->fd(bio->wio), (unsigned long) len);
+		g_warning("readv(fd=%d, len=%zu) returned -1 with errno = 0, "
+			"assuming EAGAIN", bio->wio->fd(bio->wio), len);
 		errno = VAL_EAGAIN;
 	}
